app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyjwt==2.8.0
pydantic==2.5.2
pytest==7.4.3
requests==2.31.0
httpx==0.25.1